remote-user: Fix .conf migration in %post
If the remote-user auth plug-in is installed but is not enabled at the
time of an upgrade, its httpd configuration file may not be present.  In
order to avoid causing an error during a package upgrade, only modify
the file if it is present.

This commit fixes bug 1001459.

(cherry picked from commit d2e35f26571065963f2294d2b5a0881059ce9e0a)