move_gear: drop outdated rollback code
Don't call the remove-httpd-proxy hook when rolling back a failed
move_gear operation in MCollectiveApplicationContainerProxy.

The remove-httpd-proxy hook was removed from the v2 cartridge model, so
calling the hook results in calling an undefined method in the cartridge
model when v2 cartridges are used.  Furthermore, the platform now takes
care of configuring the port proxy, so it is no longer necessary for
move_gear to invoke remove-httpd-proxy.  Commit
7d458f76116a0b72b651155d1178ad3d70398104 already dropped the call in
move_gear to the "move" hook, which invoked the deploy-httpd-proxy hook,
which is also handled by the platform now.

This commit fixes Bug 1003505.

(cherry picked from commit 5e8976026d7533ce798088399fd3c318492bfae5)