postgresql cart: Handle reload when stopped
If the postgresql cartridge is stopped in a scalable app and the user
tries to reload it using the postgresql cartridge's control script,
return an error message indicating that this is an invalid action.

Handle force-reload the same way as reload in the postgresql cartridge's
control script.

Previously, the control script would try to reload postgresql
even if it wasn't already running, which would fail.  It is preferable
to tell the user upfront that the action is invalid.

This commit closes bug 1010846.

(cherry picked from commits ee8816e280c1aeead2351943cce7b8d2e3884f06,
3273645be4a4455e733f629369a933a12a127c76,
187bae81024adcb4bd53b7334bc126152b7626d3, and
c2e9b8176b527c27caaf63a9a76177fa9bddbd58)