oo-accept-broker, oo-register-dns: unauth'd DNS
oo-accept-broker: Use the same logic as the openshift-origin-dns-nsupdate
plug-in to determine which authentication mechanism is used, and correctly
test DNS using unauthenticated updates if the plug-in is so configured.

oo-register-dns: Attempt an unauthenticated update if the user does not
indicate to use Kerberos or a keyfile and there is no keyfile at the
default location.

(cherry picked from commit 59aa27cadfda82a894bf70b8296cb832f0fffce5)